+++ index.commonjs.js
@@ -1,7 +1,4 @@
 'use strict'
 
-exports = module.exports = require('./commonjs/WithTooltip').default
-
-exports.TimeAgo = require('./commonjs/ReactTimeAgo').default
-
-exports['default'] = exports
+exports = module.exports = require('./commonjs/ReactTimeAgo').default
+exports['default'] = require('./commonjs/ReactTimeAgo').default