+++ index.js
@@ -1,2 +1,1 @@
-export { default as TimeAgo } from './modules/ReactTimeAgo'
-export { default as default } from './modules/WithTooltip'
+export { default as default } from './modules/ReactTimeAgo'