+++ test/exports.test.js
@@ -1,9 +1,6 @@
-import
-{
-	default as WithTooltip,
-	TimeAgo as ReactTimeAgo
-}
-from '../index'
+import {
+	default as ReactTimeAgo,
+} from '../index'
 
 import javascript_time_ago from 'javascript-time-ago'
 javascript_time_ago.locale(require('javascript-time-ago/locale/en'))
@@ -12,16 +9,14 @@ describe(`exports`, function()
 {
 	it(`should export ES6`, function()
 	{
-		new ReactTimeAgo({ locale: 'en-US' }, {})
-		WithTooltip({ locale: 'en-US', children: new Date() })
+		new ReactTimeAgo({ locale: 'en-US', children: new Date() })
 	})
 
 	it(`should export CommonJS`, function()
 	{
 		const Library = require('../index.commonjs')
 
-		new Library.TimeAgo({ locale: 'en-US' }, {})
-		Library.default({ locale: 'en-US', children: new Date() })
-		Library({ locale: 'en-US', children: new Date() })
+		new Library({ locale: 'en-US', children: new Date() })
+		new Library.default({ locale: 'en-US', children: new Date() })
 	})
 })