+++ tooltip/index.commonjs.js
@@ -1,0 +1,4 @@
+'use strict'
+
+exports = module.exports = require('../commonjs/ReactTimeAgoWithTooltip').default
+exports['default'] = require('../commonjs/ReactTimeAgoWithTooltip').default